4252: move getShowSealDocketRecordEntry test into separate file, as other formattedDocketEntries tests are. [skip ci]
